Create mainpage component and move fetching
